Running the powerDOREE script
This was done by adding a potentiometer to the circuit and changing its values while the batteries voltages were being read. This allowed to quickly check if the warning and ready levels would be reached and successfully trigger a signal sent. Will need further testing once integration is in place.